Refactor loading fixtures from jar resources
